Issue #961 Move Builder type to API
* Enables CacheManager.createCache to be seeded with a Builder of
CacheConfiguration
* Took the opportunity to add some more overloaded methods taking
builders when there was a version taking a configuration
