+++ impl/src/main/java/org/ehcache/config/builders/CacheManagerBuilder.java
@@ -18,6 +18,7 @@ package org.ehcache.config.builders;
 
 import org.ehcache.CacheManager;
 import org.ehcache.PersistentCacheManager;
+import org.ehcache.config.Builder;
 import org.ehcache.config.CacheConfiguration;
 import org.ehcache.config.Configuration;
 import org.ehcache.config.units.MemoryUnit;
